ai-model/requirements.txt
flask==2.3.2
scikit-learn==1.4.2
pandas==2.2.3
numpy==1.26.4
nltk==3.8.1
spacy==3.6.0
transformers==4.30.2
torch==2.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pymongo==4.4.0
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
joblib==1.3.1
